Added Sphinx docs and logo for RxPY
Sorry for the big commit, I'm not very good with Git yet.
I'm trying to learn it, though.

Sphinx docs
-----------

There is still very little content.
The format is heavily inspired by the reactivex.io website
The docs are mostly automatically generated from the docstrings
following reactivex.io's organization (similar operators are
also grouped in the same file)

I added images from ReactiveX to the operator's documentation,
using a script (actually a Jupyter notebook - 'utils.ipynb')
that tried to match filenames to operator names.
There might be mistakes.

The .rst files for the operators are meant to be used as a skeleton
over which to write the real docs.
They start with a ReST comment that marks them as automatically generated.
When someone edits them manually, the comment should be removed
so that the changes are not destroyed if the 'utils.ipynb' notebook
is run again.

The alphabetical list of operators is lso generated automatically,
and shouldn't be edited manually.

Logo
----

I thought RxPY should have a logo.
It is based on ReactiveX's purple dragon, turned into
two snakes (like Python's logo) and rendered with Python's colors.

I also created a favicon, based on a stilized version of the logo.