Don't override custom exceptions in throw.py
Say I have a custom exception:

    class TimeoutException(Exception):
        pass

I throw this if some observable times out using using

    observable.timeout(1000, Observable.throw(TimeoutError))

so that I can later switch on the type of the exception in a .catch_exception call.

The file change allows this behaviour. The old behaviour was to override the exception type to simply Exception (since type(TimeoutException) != Exception).