Refactoring test_catch pt 2 (1/2): Updating test boilerplate to account for using handlers.
We lose a little bit of coverage regarding checking that handlers see
the errors that we expect them to see. This could be re-introduced by
taking a list of handlers to use instead of one handler-builder used to
create all handlers.

I think this loss of coverage is worth more declarative tests.