implemented threadsafe schedule methods (#266)
This commit fixes #249.

The immediate scheduling method is based on call_soon_threadsafe to
schedule the action on the event loop. The dispose function uses
a future from the concurrent package so that it can block until the
action is canceled.

The relative scheduling method is executed in two steps: The first one
allows to move the execution context to the event loop. The second step
schedules the action at the appropriate time.