Fixed await on Observable that schedules later
In the current implementation, tofuture does not provide a scheduler
when subscribing to the Observable. As a consequence, when used from
await,  if the Observable chain uses an operator that deals with time,
a deadlock occurs because we try to complete the - asyncio - future
from another thread than the mainloop.

In __await__ we can provide an AsyncioScheduler associated to the
current event loop because it is always called from the context of the
current event loop.

Fixes #456