Fix collections imports for Python 3.10+
Summary: Right now RxPY 1.6.1 imports `Iterable` from `collections`,
which is deprecated and will break in Python 3.10. This is a PR for the
release/v1.6.x branch which fixes that (in a backwards compatible way),
so that projects stuck with RxPY 1.6.x can still upgrade to Python
3.10. My hope is that this can be released as a new minor version
(probably 1.6.2) of RxPY.

For context, we have a project that depends on RxPY via another
dependency. Unfortunately, that dependency only supports RxPY <2, and
upgrading that seems non-trivial.

Test Plan: I ran `python -m unittest discover tests` before and after
this change with both Python 3.9 and Python 2.7. I confirmed the number
of tests passed before and after this change was the same (though with
2.7 there were 17 errors even before this change, and with 3.9 there
were 17 errors and 5 failures even before this change. Not sure if
that is due to any configuration issues on my end).